feat: 출석 여부 검증 코드 추가

+++ src/main/java/katecam/luvicookie/ditto/domain/attendance/application/AttendanceService.java
@@ -127,6 +127,11 @@ public class AttendanceService {
             return;
         }
 
+        // 출석 여부 확인
+        if (!attendanceRepository.existsByAttendanceDate_IdAndMember_Id(attendanceDate.getId(), memberId)) {
+            throw new GlobalException(ErrorCode.NOT_ATTENDED);
+        }
+
         attendanceRepository.delete(attendance);
     }
